Add element above router
